+++ CotEditor/Sources/FileEncoding.swift
@@ -8,7 +8,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2020-2022 1024jp
+//  © 2020-2024 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -23,7 +23,7 @@
 //  limitations under the License.
 //
 
-struct FileEncoding: Equatable {
+struct FileEncoding: Equatable, Hashable {
     
     var encoding: String.Encoding
     var withUTF8BOM: Bool = false